initial

+++ src/main/java/group12/sebm/controller/dto/UserDto.java
@@ -1,21 +1,21 @@
 package group12.sebm.controller.dto;
 
 public class UserDto {
-    private Long id;
+    private Integer id;
     private String username;
     private String password;
 
-    public UserDto(Long id, String username, String password) {
+    public UserDto(Integer id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
